ajout des modules angular pour facilite la maintenance refactorisation du code suppression des compononent dans le fichier app.modules de l'application est mise a jour du style de la nav bar
