+++ package.json
@@ -10,11 +10,13 @@
   },
   "private": true,
   "dependencies": {
+    "@angular-material-components/file-input": "^8.0.0",
     "@angular/animations": "^14.0.0",
     "@angular/cdk": "^14.1.2",
     "@angular/common": "^14.0.0",
     "@angular/compiler": "^14.0.0",
     "@angular/core": "^14.0.0",
+    "@angular/elements": "^14.1.3",
     "@angular/forms": "^14.0.0",
     "@angular/material": "^14.1.2",
     "@angular/platform-browser": "^14.0.0",
