+++ src/app/core/core.module.ts
@@ -1,0 +1,15 @@
+
+
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@NgModule({
+  declarations: [
+  ],
+  imports: [
+    CommonModule,
+  ],
+  exports: [
+  ]
+})
+export class CoreModule { }
