+++ src/app/core/models/groupamania-post.model.ts
@@ -4,10 +4,10 @@ export class GroupamaniaGeneralPost {
 
   constructor(
     public _id: string,
-    public FirstName: string,
+    // public FirstName: string,
     public post: string,
-    public LastName: string,
-    public PostGroupamania: string,
+    // public LastName: string,
+    // public PostGroupamania: string,
     public ImageUrl: string,
     public userIdPoster: string,
   ) {
