+++ src/app/core/service/groupamania.service.spec.ts
