--- src/app/general-post/general-post.component.scss
@@ -1,16 +1,0 @@
-.mat-card {
-    width: 50%;
-}
-
-.container {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-top: 5%;
-    &-card {
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-    }
-}
