--- src/app/general-post/general-post.component.ts
@@ -1,36 +1,0 @@
-import { Router } from '@angular/router';
-
-import { Component, Input, OnInit } from '@angular/core';
-import { GroupamaniaGeneralPost } from '../models/groupamania-post.model';
-import { UserToken } from '../models/user-token.model';
-import { LocalizedString, Token } from '@angular/compiler';
-
-
-@Component({
-  selector: 'app-general-post',
-  templateUrl: './general-post.component.html',
-  styleUrls: ['./general-post.component.scss']
-})
-export class GeneralPostComponent implements OnInit {
-  tokenLocalStorage!: any;
-  userId!: any;
-
-
-  @Input() post!: GroupamaniaGeneralPost;
-
-
-  constructor(private router: Router) { }
-
-  ngOnInit(): void {
-    this.tokenLocalStorage = localStorage.getItem("access_token");
-    this.userId = localStorage.getItem("ID");
-
-  }
-
-
-  onPostById() {
-    this.router.navigateByUrl("accueille/" + this.post._id);
-    return this.post._id
-  }
-
-}
