--- src/app/groupamania/groupamania.component.html
@@ -1,1 +1,0 @@
-<app-general-post *ngFor="let Post of Posts$ | async" [post]="Post"></app-general-post>
