--- src/app/groupamania/groupamania.component.scss
@@ -1,3 +1,0 @@
-.container_post {
-    margin-top: 10%;
-}
