--- src/app/nav-bar/nav-bar.component.html
@@ -1,15 +1,0 @@
-<nav>
-    <ul>
-        <li class="title">Groupamania</li>
-        <li (click)="onAcceuille()">Accueille
-            <fa-icon [icon]="faHouse"></fa-icon>
-        </li>
-        <li (click)="onProfil()">
-            Profil
-            <fa-icon [icon]="faUser"></fa-icon>
-        </li>
-        <li (click)="onDeconnexion()" class="logout_button">Deconnexion</li>
-    </ul>
-
-
-</nav>
