--- src/app/nav-bar/nav-bar.component.scss
@@ -1,41 +1,0 @@
-@import "src/app/assets/scss/_variable.scss";
-nav {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    font-size: 2rem;
-    color: $tertiaire;
-}
-
-@font-face {
-    font-family: "DancingScript";
-    src: url('../../assets/font/DancingScript-VariableFont_wght.ttf') format("opentype");
-}
-
-ul {
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-    overflow: hidden;
-}
-
-.title {
-    font-family: "DancingScript";
-    margin-right: 5%;
-    font-size: 2rem;
-    width: 25%;
-}
-
-li {
-    float: left;
-    margin: 0 2% 0 2%;
-    &:hover {
-        cursor: pointer;
-        color: $primaire;
-    }
-}
-
-.logout_button {
-    float: right;
-}
