--- src/app/one-post/one-post.component.html
@@ -1,15 +1,0 @@
-<app-nav-bar></app-nav-bar>
-<div class="container">
-    <mat-card>
-        <div class="container-card">
-            <img mat-card-image [src]="" [alt]="">
-            <mat-card-content>
-                {{this.thePost}}
-            </mat-card-content>
-            <mat-card-actions>
-                <button mat-raised-button color="primary" type="submit" *ngIf="this.IdUser === this.userIdPoster">Modifier le post</button>
-                <button mat-raised-button color="primary" type="submit" *ngIf="this.IdUser === this.userIdPoster">Supprimer le post</button>
-            </mat-card-actions>
-        </div>
-    </mat-card>
-</div>
