--- src/app/page-accueille/page-accueille.component.html
@@ -1,2 +1,0 @@
-<app-nav-bar></app-nav-bar>
-<app-groupamania></app-groupamania>
