--- src/app/page-accueille/page-accueille.component.scss
@@ -1,35 +1,0 @@
-// .container_accueille {
-//     margin-top: 10%;
-// }
-// .card_for_post {
-//     width: 60%;
-//     border: 1px solid black;
-//     display: flex;
-//     flex-direction: column;
-//     align-items: center;
-//     & .container_commentaire {
-//         display: flex;
-//         align-items: center;
-//         & img {
-//             margin-right: 2%;
-//             width: 5%;
-//             border-radius: 100%;
-//         }
-//         & input {
-//             height: 1.5rem;
-//             width: 80%;
-//             border-radius: 1.5rem;
-//             outline: none;
-//         }
-//         & .for_commentaire {
-//             border: 1px solid black;
-//             width: 100%;
-//         }
-//     }
-// }
-// ul {
-//     list-style-type: none;
-// }
-// li:hover {
-//     border: 1px solid black;
-// }
