--- src/app/page-accueille/page-accueille.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PageAccueilleComponent } from './page-accueille.component';
-
-describe('PageAccueilleComponent', () => {
-  let component: PageAccueilleComponent;
-  let fixture: ComponentFixture<PageAccueilleComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ PageAccueilleComponent ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(PageAccueilleComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
