--- src/app/page-accueille/page-accueille.component.ts
@@ -1,18 +1,0 @@
-
-import { Component, Input, OnInit } from '@angular/core';
-
-
-@Component({
-  selector: 'app-page-accueille',
-  templateUrl: './page-accueille.component.html',
-  styleUrls: ['./page-accueille.component.scss']
-})
-export class PageAccueilleComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-
-  }
-
-}
