--- src/app/page-profil/page-profil.component.html
@@ -1,8 +1,0 @@
-<style>
-    p {
-        margin: 20%;
-    }
-</style>
-
-<app-nav-bar></app-nav-bar>
-<p>page-profil works!</p>
