--- src/app/page-profil/page-profil.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-page-profil',
-  templateUrl: './page-profil.component.html',
-  styleUrls: ['./page-profil.component.scss']
-})
-export class PageProfilComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
