--- src/app/redirecting-connexion/redirecting-connexion.component.html
@@ -1,33 +1,0 @@
-<div class="container-formulaire">
-    <form [formGroup]="addressForm" novalidate>
-        <mat-card class="shipping-card">
-            <mat-card-header>
-                <mat-card-title>Connexion</mat-card-title>
-            </mat-card-header>
-            <mat-card-content>
-                <div class="row">
-                    <div class="col">
-                        <mat-form-field class="full-width">
-                            <input matInput placeholder="First email" formControlName="email">
-                            <mat-error *ngIf="addressForm.controls['email'].hasError('required')">
-                                First name is <strong>required</strong>
-                            </mat-error>
-                        </mat-form-field>
-                    </div>
-                    <div class="col">
-                        <mat-form-field class="full-width">
-                            <input matInput placeholder="Last name" type="password" formControlName="password">
-                            <mat-error *ngIf="addressForm.controls['password'].hasError('required')">
-                                Last name is <strong>required</strong>
-                            </mat-error>
-                        </mat-form-field>
-                    </div>
-                </div>
-
-            </mat-card-content>
-            <mat-card-actions>
-                <button (click)="onConnexion()" mat-raised-button color="primary" type="submit">Submit</button>
-            </mat-card-actions>
-        </mat-card>
-    </form>
-</div>
