--- src/app/redirecting-connexion/redirecting-connexion.component.scss
@@ -1,17 +1,0 @@
-.mat-card {
-    width: 100%;
-    height: auto;
-    margin-top: 5%;
-}
-
-input {
-    font-size: 15px;
-}
-
-.container {
-    &-formulaire {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-}
