+++ src/app/router/router.module.ts
@@ -3,21 +3,22 @@
 
 import { PageNotFoundComponent } from './../page-not-found/page-not-found.component';
 
-import { OnePostComponent } from './../one-post/one-post.component';
+import { OnePostComponent } from '../groupamania-module/components/one-post/one-post.component';
 import { NgModule, Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule , Routes , Router } from '@angular/router';
 
-import { PageAccueilleComponent } from '../page-accueille/page-accueille.component';
-import { PageProfilComponent } from '../page-profil/page-profil.component';
-import { RedirectingConnexionComponent } from './../redirecting-connexion/redirecting-connexion.component';
+import { PageAccueilleComponent } from '../groupamania-module/components/page-accueille/page-accueille.component';
+import { PageProfilComponent } from '../groupamania-module/components/page-profil/page-profil.component';
+import { SignInLoginComponent } from '../groupamania-module/components/sign-in-login/sign-in-login.component';
+
 
 
 
 const routes: Routes = [
   { path: 'accueille', component: PageAccueilleComponent },
   { path: 'profil', component: PageProfilComponent },
-  { path: "", component: RedirectingConnexionComponent },
+  { path: "", component: SignInLoginComponent },
   { path: 'accueille/:id', component: OnePostComponent },
   { path: "**" , component: PageNotFoundComponent}
 ];
